A game called PIG
This a a jeopardy style game where the goal is to be the first person to to reach a score of 100.  The player will roll a die and if it lands on 1, they loose all their points from that turn. Otherwise, the dice roll is added to their score. The player can then roll again or hold and end their turn.